+++ lib/couchdb.dart
@@ -34,7 +34,7 @@ export 'src/documents.dart';
 export 'src/local_documents.dart';
 export 'src/server.dart';
 
-@Deprecated("Use classes above instead of the ones below")
+@Deprecated('Use classes above instead of the ones below')
 export 'src/entities/db_response.dart';
 export 'src/entities/database_model_response.dart';
 export 'src/entities/design_document_model_response.dart';