+++ lib/src/models/database_model.dart
@@ -1,7 +1,7 @@
 import 'package:couchdb/couchdb.dart';
 
 /// Only provided for backward compatibility
-@Deprecated("Use Databases innstead")
+@Deprecated('Use Databases instead')
 class DatabaseModel extends Databases {
   DatabaseModel(ClientInterface client) : super(client);
 }