+++ lib/src/models/design_document_model.dart
@@ -1,7 +1,7 @@
 import 'package:couchdb/couchdb.dart';
 
 /// Only provided for backward compatibility
-@Deprecated("Use DesignDocuments innstead")
+@Deprecated('Use DesignDocuments instead')
 class DesignDocumentModel extends DesignDocuments {
   DesignDocumentModel(ClientInterface client) : super(client);
 }