+++ lib/src/models/document_model.dart
@@ -1,7 +1,7 @@
 import 'package:couchdb/couchdb.dart';
 
 /// Only provided for backward compatibility
-@Deprecated("Use Documents innstead")
+@Deprecated('Use Documents instead')
 class DocumentModel extends Documents {
   DocumentModel(ClientInterface client) : super(client);
 }