+++ lib/src/models/local_document_model.dart
@@ -1,7 +1,7 @@
 import 'package:couchdb/couchdb.dart';
 
 /// Only provided for backward compatibility
-@Deprecated("Use LocalDocuments innstead")
+@Deprecated('Use LocalDocuments instead')
 class LocalDocumentModel extends LocalDocuments {
   LocalDocumentModel(ClientInterface client) : super(client);
 }