+++ test/couchdb_test.dart
@@ -1,7 +1,7 @@
 import 'package:couchdb/couchdb.dart';
 
 Future<void> main() async {
-  final c = CouchDbClient.fromString('http://localhost:5984');
+  final c = CouchDbClient.fromString('http://admin:password@127.0.0.1:5984');
   final da = Databases(c);
   final ddm = DesignDocuments(c);
   final dm = Documents(c);
@@ -11,11 +11,8 @@ Future<void> main() async {
 
   try {
     final headers = <String, String>{'Accept': 'text/plain'};
-
-    //final r = await c.authenticate();
-
-    final o = await da.shards('denta');
-    print(o.shards);
+    final o = await da.dbInfo('denta');
+    print(o.dbName);
   } on CouchDbException catch (e) {
     print(e);
   }